Added more errors checks to generateInput.sh.
generateInput.sh now checks that all necessary files have made it into the output text file and that the MISR/MODIS file lists start with the correct file. Also modified loadModules so that the module commands are verbose. Added example errors to the inputFileDebug (the functionality added to generateInput.sh is represented there).