Updated README, makefile, MODIS and libTERRA
Updated the README to contain more up to date steps for setting up the program.
More specifically, included better steps to set up the Makefile.

Added the first MODIS dimension scale code. Next step is to turn that code into
a function and apply it to all of the MODIS datasets.

Added to libTERRA the change_dim_attr_NAME_value function.