MOPITT subsetting complete
The MOPITT code appears to subset correctly based on the orbit struct passed to it. Still a lot of unknowns about if the subsetting is precisely correct, but the binary search I have implemented always returns the FIRST element greater than the target value (the target value would be given by the orbit start and end times).