CERES has been updated for the new file
CERES has been rewritten to include the new SSF file and datasets. The code is much neater than before as well. CERES code is now ready for subsetting if needed, simply replace the readThenWrite function calls with a new CERESreadThenWrite.

Also updated the existing readThenWrite function to include the ability to specify the name of the output HDF5 dataset during function calls. Please see the comments for that function in libTERRA.c.