MOPITT dual granule handling complete
This commit adds the ability for MOPITT to handle at most two granules. One ntrack dimension have been added for each
granule added to the file.

Also fixed issue with jobSubmit not properly modifying the batchscript.pbs file to contain the correct input file list for
the fusion program.

Fixed the CERES infinite loop problem that was causing an infinite number of granules to be added to the output file. This was discovered because the program ran past its allotted wallclock time and was terminated by Blue Waters.