Master script for tar files
This commit adds the master script for tarring the BF files on ROGER (tarAndTransfer.sh). Also included are various improvements and bug fixes to python scripts. The Globus Python command line interface was added to the configureEnv.sh configuration script so that the project's python virtual environment now has access to Globus CLI.